+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Export valid empty JSON instead of invalid empty string
+
 ## [1.1.0] - 2025-01-14
 
 This version has been generated using: